requirements.txt
python-telegram-bot>=21.0
requests>=2.31.0
python-dotenv>=1.0.0
psutil>=5.9.0
scapy>=2.5.0
httpx>=0.25.0
dnspython>=2.4.0
